% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rimdb.R
\name{imdb_criar_query}
\alias{imdb_criar_query}
\title{Função que cria a query de pesquisa}
\usage{
imdb_criar_query(q, t = "e")
}
\arguments{
\item{q}{palavras chave pesquisadas}

\item{t}{tipo da pesquisa ("e" pesquisa por titulo extao, "ne" pode fazer aproximações)}
}
\description{
Função que cria a query de pesquisa
}

