% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rimdb.R
\name{imdb_infos_tt}
\alias{imdb_infos_tt}
\title{Retorna as infos de um filme (usando o tt)}
\usage{
imdb_infos_tt(tt)
}
\arguments{
\item{tt}{codigo de um filme no imdb}
}
\description{
Retorna as infos de um filme (usando o tt)
}

