% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rimdb.R
\name{imdb_parse_pesquisa}
\alias{imdb_parse_pesquisa}
\title{Retorna lista com resultado das palavras pesquisadas}
\usage{
imdb_parse_pesquisa(p)
}
\arguments{
\item{p}{pesquisa retornada pela função imdb_pesquisa}
}
\description{
Retorna lista com resultado das palavras pesquisadas
}

