% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rimdb.R
\name{imdb_pesquisa}
\alias{imdb_pesquisa}
\title{Função que retorna o resultado da pesquisa (formato html)}
\usage{
imdb_pesquisa(q, t = "e")
}
\arguments{
\item{q}{palavras chave pesquisadas}

\item{t}{tipo da pesquisa ("e" pesquisa por titulo extao, "ne" pode fazer aproximações)}
}
\description{
Função que retorna o resultado da pesquisa (formato html)
}

